pyproject.toml
[project]
name = "karo"
version = "0.1.0"
description = "AI agent that ingests internal knowledge and lets anyone query live data in plain English"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.3.0",
    "faster-whisper>=1.1.0",
    "langchain>=1.2.9",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.8",
    "langchain-postgres>=0.0.16",
    "langgraph>=1.0.8",
    "langgraph-checkpoint-postgres>=3.0.4",
    "pgvector>=0.3.6",
    "psycopg2-binary>=2.9.11",
    "psycopg[binary]>=3.3.2",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.46",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "pyyaml>=6.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
# One loop per session: AsyncPostgresSaver + compiled graph are process singletons
# and must not be reused across different asyncio loops.
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
markers = [
    "agent: full agent turn (LLM + tools; costs tokens)",
    "integration: requires working API_TOKEN and live API",
]
